src/core/patterns/symplectic.py
"""Symplectic geometry utilities for pattern spaces.

This module provides tools for working with symplectic structures in pattern spaces,
including symplectic forms, Hamiltonian flows, and Poisson brackets. The implementation
follows an enriched categorical structure using operadic transitions and wave emergence
for natural dimensional changes.

Key Features:
    - Quantum geometric tensor implementation (Q_{μν} = g_{μν} + iω_{μν})
    - Wave emergence integration for smooth transitions
    - Operadic structure for dimension handling
    - Ricci flow for geometric evolution
    - Non-commutative geometry support

Mathematical Framework:
    The implementation is based on the following mathematical structures:
    
    1. Symplectic Form:
       ω: TM × TM → R satisfying:
       - Antisymmetry: ω(X,Y) = -ω(Y,X)
       - Non-degeneracy: ω(X,Y) = 0 ∀Y ⟹ X = 0
    
    2. Quantum Geometric Tensor:
       Q_{μν} = g_{μν} + iω_{μν}
       - g_{μν}: Riemannian metric
       - ω_{μν}: Symplectic form
    
    3. Wave Emergence:
       Ψ(x) = ∫ dk/√(2ω_k) (a(k)e^{-ikx} + a†(k)e^{ikx})
    
    4. Ricci Flow:
       ∂g/∂t = -2Ric(g)

Example:
    >>> structure = SymplecticStructure(dim=4, wave_enabled=True)
    >>> point = torch.randn(4)
    >>> evolved = structure.quantum_ricci_flow(point, time=1.0)
"""

from __future__ import annotations
from dataclasses import dataclass
from typing import Tuple, Optional, Dict, Any, cast, Protocol
import torch
from torch import Tensor
from .operadic_structure import (
    AttentionOperad,
    OperadicOperation,
    EnrichedAttention,
)

__all__ = ['SymplecticForm', 'SymplecticStructure']

class WaveOperator(Protocol):
    """Protocol defining wave operator interface."""
    def wave_operator(self, tensor: Tensor) -> Tensor: ...
    def create_wave_packet(self, position: Tensor, momentum: Tensor) -> Tensor: ...
    def get_position(self, wave: Tensor) -> Tensor: ...
    def get_momentum(self, wave: Tensor) -> Tensor: ...
    def create_morphism(self, pattern: Tensor, operation: OperadicOperation, include_wave: bool = True) -> Tensor: ...

class EnrichedOperator(WaveOperator, Protocol):
    """Protocol for enriched operators with wave behavior."""
    base_category: str
    wave_enabled: bool

@dataclass
class SymplecticForm:
    """Symplectic form on a manifold with enriched structure.
    
    This class implements a symplectic form with additional enriched structure data.
    The form satisfies the key properties of antisymmetry and non-degeneracy.
    
    Attributes:
        matrix: The symplectic form matrix representation
        enrichment: Optional dictionary containing enriched structure data
        
    Properties:
        - Antisymmetry is enforced in __post_init__
        - Matrix representation includes enrichment data
        - Standard operations (evaluate, transpose, negation) preserve structure
    
    Mathematical Properties:
        1. Antisymmetry: ω(X,Y) = -ω(Y,X)
        2. Non-degeneracy: ω(X,Y) = 0 ∀Y ⟹ X = 0
        3. Closed form: dω = 0
    """
    
    matrix: Tensor  # The symplectic form matrix
    enrichment: Optional[Dict[str, Any]] = None  # Enriched structure data
    
    def __post_init__(self) -> None:
        """Validate symplectic form properties."""
        if not torch.allclose(self.matrix, -self.matrix.transpose(-1, -2)):
            raise ValueError("Symplectic form must be antisymmetric")

    def evaluate(self, v1: Tensor, v2: Tensor) -> Tensor:
        """Evaluate symplectic form on two vectors with structure preservation.
        
        Args:
            v1: First vector
            v2: Second vector
            
        Returns:
            Symplectic form evaluation ω(v1,v2)
        """
        return torch.einsum('...ij,...i,...j->...', self.matrix, v1, v2)

    def transpose(self, *args) -> 'SymplecticForm':
        """Return transposed symplectic form preserving enriched structure."""
        return SymplecticForm(
            matrix=self.matrix.transpose(-2, -1),
            enrichment=self.enrichment
        )

    def __neg__(self) -> 'SymplecticForm':
        """Return negated symplectic form preserving enriched structure."""
        return SymplecticForm(
            matrix=-self.matrix,
            enrichment=self.enrichment
        )

    def __eq__(self, other: object) -> bool:
        """Check equality with another symplectic form."""
        if not isinstance(other, SymplecticForm):
            return NotImplemented
        return torch.allclose(self.matrix, other.matrix)


class SymplecticStructure:
    """Symplectic structure on a manifold with enriched operadic transitions.
    
    This class provides methods for working with symplectic geometry using
    enriched categorical structure and operadic transitions for dimension
    changes. It integrates wave emergence behavior for natural transitions.
    
    Key Features:
        1. Dimension Handling:
           - Uses operadic transitions instead of padding
           - Preserves structure through enriched morphisms
           - Handles wave emergence behavior
        
        2. Quantum Geometry:
           - Implements quantum geometric tensor Q_{μν} = g_{μν} + iω_{μν}
           - Supports non-commutative geometry
           - Includes quantum Ricci flow
        
        3. Wave Integration:
           - Wave packet creation and evolution
           - Phase space integration
           - Hamiltonian flow tracking
    
    Attributes:
        dim: Original dimension of the manifold
        target_dim: Target dimension (always even) for symplectic structure
        wave_enabled: Whether wave emergence behavior is enabled
        preserve_structure: Whether to preserve symplectic structure
        operadic: Operadic structure for transitions
        enriched: Enriched attention structure
        
    Mathematical Framework:
        1. Symplectic Form:
           ω = Σ dxᵢ ∧ dpᵢ
        
        2. Quantum Geometric Tensor:
           Q_{μν} = g_{μν} + iω_{μν}
        
        3. Wave Function:
           Ψ(x) = ∫ dk/√(2ω_k) (a(k)e^{-ikx} + a†(k)e^{ikx})
        
        4. Ricci Flow:
           ∂g/∂t = -2Ric(g)
    """
    
    # Class constants
    _SYMPLECTIC_WEIGHT: float = 0.1  # Weight for fiber coordinate terms

    def __init__(
        self,
        dim: int,
        preserve_structure: bool = True,
        wave_enabled: bool = True,
        dtype: torch.dtype = torch.float32
    ):
        """Initialize symplectic structure with enriched features.
        
        Args:
            dim: Dimension of the manifold (can be odd or even)
            preserve_structure: Whether to preserve symplectic structure
            wave_enabled: Whether to enable wave emergence behavior
            dtype: Data type for tensors
            
        Raises:
            ValueError: If dimension is less than 2
        """
        if dim < 2:
            raise ValueError(f"Dimension must be at least 2, got {dim}")
            
        self.dim = dim
        self.preserve_structure = preserve_structure
        self.wave_enabled = wave_enabled
        self.dtype = dtype
        
        # Target dimension is always even to maintain symplectic properties
        self.target_dim = dim if dim % 2 == 0 else dim + 1
        
        # Initialize enriched operadic structure
        self.operadic = AttentionOperad(
            base_dim=dim,
            preserve_symplectic=preserve_structure,
            preserve_metric=True,
            dtype=dtype
        )
        
        # Initialize enriched attention structure
        self.enriched = EnrichedAttention(
            base_category="SymplecticVect",
            wave_enabled=wave_enabled,
            dtype=dtype
        )
        self.enriched.wave_enabled = wave_enabled

    def _handle_dimension(self, tensor: Tensor, recursion_depth: int = 0) -> Tensor:
        """Handle dimension transitions using enriched operadic structure.
        
        Instead of padding, we use operadic transitions with wave emergence
        to naturally handle dimensional changes through enriched morphisms.
        
        Args:
            tensor: Input tensor to transform
            recursion_depth: Current recursion depth (to prevent infinite recursion)
            
        Returns:
            Transformed tensor with correct symplectic dimensions
            
        Raises:
            ValueError: If tensor dimension is invalid or operation fails
            RuntimeError: If transformation fails or recursion limit exceeded
        """
        if not isinstance(tensor, Tensor):
            raise ValueError(f"Expected torch.Tensor, got {type(tensor)}")
            
        if tensor.dim() < 1:
            raise ValueError(f"Tensor must have at least 1 dimension, got {tensor.dim()}")
            
        current_dim = tensor.shape[-1]
        if current_dim == self.target_dim:
            return tensor
            
        if current_dim < 1:
            raise ValueError(f"Last dimension must be at least 1, got {current_dim}")

        # Prevent infinite recursion
        if recursion_depth > 10:  # Maximum recursion depth
            raise RuntimeError("Maximum recursion depth exceeded in dimension handling")

        try:
            # Save original shape for proper reshaping
            original_shape = tensor.shape[:-1]
            
            # Simple dimension adjustment if wave emergence is disabled
            if not self.wave_enabled:
                if current_dim < self.target_dim:
                    # Pad with zeros
                    padding_size = self.target_dim - current_dim
                    padding = torch.zeros(*original_shape, padding_size, device=tensor.device, dtype=tensor.dtype)
                    return torch.cat([tensor, padding], dim=-1)
                else:
                    # Truncate
                    return tensor[..., :self.target_dim]

            # Use operadic transition with wave emergence
            operation = self.operadic.create_transition(current_dim, self.target_dim)
            transformed = self.enriched.create_morphism(
                tensor, 
                operation,
                include_wave=self.wave_enabled
            )
            
            # Verify the transformation worked
            if transformed.shape[-1] != self.target_dim:
                # Try one more time with increased recursion depth
                return self._handle_dimension(transformed, recursion_depth + 1)
                
            return transformed

        except Exception as e:
            raise RuntimeError(f"Failed to transform tensor dimensions: {str(e)}")

    def compute_metric(self, fiber_coords: Tensor) -> Tensor:
        """Compute Riemannian metric tensor at given fiber coordinates.
        
        Args:
            fiber_coords: Fiber coordinates to compute metric at
            
        Returns:
            Metric tensor with proper structure
        """
        # Get dimension for metric computation
        dim = fiber_coords.shape[-1]
        
        # Create base metric tensor
        metric = torch.eye(dim, device=fiber_coords.device, dtype=fiber_coords.dtype)
        
        # Add fiber coordinate dependence
        # This ensures the metric varies smoothly over the fiber
        coord_norm = torch.norm(fiber_coords, dim=-1, keepdim=True)
        scale_factor = torch.exp(-self._SYMPLECTIC_WEIGHT * coord_norm)
        metric = metric * scale_factor.unsqueeze(-1)
        
        return metric

    def standard_form(self, device: Optional[torch.device] = None) -> SymplecticForm:
        """Compute standard symplectic form with enriched structure.
        
        Args:
            device: Optional device to place tensor on
            
        Returns:
            Standard symplectic form with enriched structure
        """
        n = self.target_dim // 2  # Always even dimension
        omega = torch.zeros(self.target_dim, self.target_dim, device=device)
        omega[:n, n:] = torch.eye(n, device=device)
        omega[n:, :n] = -torch.eye(n, device=device)
        
        # Add enriched structure data
        enrichment = {
            'dimension': self.target_dim,
            'structure_type': 'standard_symplectic',
            'preserve_symplectic': True,
            'wave_enabled': self.wave_enabled
        }
        
        return SymplecticForm(omega, enrichment=enrichment)

    def compute_form(self, fiber_coords: Tensor) -> SymplecticForm:
        """Compute symplectic form for given fiber coordinates.
        
        Args:
            fiber_coords: Fiber coordinates to compute form at
            
        Returns:
            Symplectic form with enriched structure
            
        Raises:
            ValueError: If fiber coordinates have invalid dimensions
            RuntimeError: If form computation fails
        """
        try:
            # Handle dimension transition if needed
            fiber_coords = self._handle_dimension(fiber_coords)
            
            # Get original dimension before padding
            original_dim = min(self.dim, fiber_coords.shape[-1])
            n = original_dim // 2  # Use original dimension for form size
            
            # Create standard form matrix
            omega = torch.zeros(
                original_dim,
                original_dim,
                device=fiber_coords.device,
                dtype=fiber_coords.dtype
            )
            
            if n > 0:  # Only fill if we have pairs to work with
                # Standard symplectic form in block form
                # [ 0   I ]
                # [-I   0 ]
                omega[:n, n:] = torch.eye(n, device=fiber_coords.device)
                omega[n:, :n] = -torch.eye(n, device=fiber_coords.device)
                
                # Scale the form to ensure eigenvalues are large enough
                omega *= 2.0
            
            # Add enriched structure data
            enrichment = {
                'dimension': original_dim,
                'structure_type': 'standard_symplectic',
                'preserve_symplectic': self.preserve_structure,
                'wave_enabled': self.wave_enabled,
                'fiber_coords': fiber_coords[..., :original_dim]
            }
            
            return SymplecticForm(omega, enrichment=enrichment)
            
        except Exception as e:
            raise RuntimeError(f"Failed to compute symplectic form: {str(e)}") from e

    def compute_volume(self, point: Tensor) -> Tensor:
        """Compute symplectic volume form using enriched structure.
        
        Args:
            point: Point on manifold
            
        Returns:
            Volume form value with structure preservation
        """
        point_symplectic = self._handle_dimension(point)
        form = self.compute_form(point_symplectic)
        # Volume is Pfaffian of symplectic form
        # For standard form with enriched structure, this is 1
        return torch.ones(1, device=point_symplectic.device)

    def hamiltonian_vector_field(
        self,
        hamiltonian: Tensor,
        point: Tensor
    ) -> Tensor:
        """Compute Hamiltonian vector field using enriched structure.
        
        Args:
            hamiltonian: Hamiltonian function value
            point: Point on manifold
            
        Returns:
            Hamiltonian vector field preserving enriched structure
        """
        # Handle dimensional transition using enriched structure
        point_symplectic = self._handle_dimension(point)
        form = self.compute_form(point_symplectic)
        grad_h = cast(Tensor, torch.autograd.grad(hamiltonian, point_symplectic, create_graph=True)[0])
        
        # Compute field in symplectic dimension with structure preservation
        field = torch.einsum('ij,j->i', form.matrix, grad_h)
        
        # Project back to original dimension if needed
        if field.shape[-1] != self.dim:
            # Create enriched operation for projection
            operation = self.operadic.create_operation(
                source_dim=field.shape[-1],
                target_dim=self.dim,
                preserve_structure='symplectic'
            )
            
            # Project through enriched structure with wave behavior
            field = self.enriched.create_morphism(
                pattern=field.reshape(-1, field.shape[-1]),
                operation=operation,
                include_wave=True
            )
        
        return field

    def poisson_bracket(
        self,
        f: Tensor,
        g: Tensor,
        point: Tensor,
    ) -> Tensor:
        """Compute Poisson bracket using enriched structure.
        
        Args:
            f: First function value
            g: Second function value
            point: Point on manifold
            
        Returns:
            Poisson bracket value preserving enriched structure
        """
        point_symplectic = self._handle_dimension(point)
        form = self.compute_form(point_symplectic)
        grad_f = cast(Tensor, torch.autograd.grad(f, point_symplectic, create_graph=True)[0])
        grad_g = cast(Tensor, torch.autograd.grad(g, point_symplectic, create_graph=True)[0])
        return form.evaluate(grad_f, grad_g) 

    def compute_quantum_geometric_tensor(self, point: Tensor) -> Tensor:
        """Compute quantum geometric tensor Q_{μν} = g_{μν} + iω_{μν}.
        
        Args:
            point: Point on manifold to compute tensor at
            
        Returns:
            Complex tensor representing quantum geometry
        """
        # Compute metric and symplectic components
        metric = self.compute_metric(point)
        symplectic = self.compute_form(point).matrix
        
        # Combine into quantum geometric tensor
        return metric + 1j * symplectic

    def compute_ricci_tensor(self, point: Tensor) -> Tensor:
        """Compute Ricci tensor for quantum geometry.
        
        Args:
            point: Point on manifold
            
        Returns:
            Ricci tensor at given point
        """
        # Get quantum geometric tensor
        Q = self.compute_quantum_geometric_tensor(point)
        
        # Extract metric and symplectic parts
        g = Q.real
        omega = Q.imag
        
        # Compute Christoffel symbols
        n = g.shape[-1]
        gamma = torch.zeros(n, n, n, device=point.device, dtype=point.dtype)
        
        # First compute partial derivatives of metric
        # For simplicity, we use finite differences
        eps = 1e-6
        for k in range(n):
            point_plus = point.clone()
            point_plus[..., k] += eps
            point_minus = point.clone()
            point_minus[..., k] -= eps
            
            g_plus = self.compute_metric(point_plus)
            g_minus = self.compute_metric(point_minus)
            
            dg = (g_plus - g_minus) / (2 * eps)
            
            for i in range(n):
                for j in range(n):
                    for l in range(n):
                        gamma[i,j,k] += 0.5 * g[i,l] * (
                            dg[l,j] + dg[j,l] - dg[j,l]
                        )
        
        # Compute Riemann tensor
        R = torch.zeros(n, n, n, n, device=point.device, dtype=point.dtype)
        for i in range(n):
            for j in range(n):
                for k in range(n):
                    for l in range(n):
                        # R^i_{jkl} = ∂_k Γ^i_{jl} - ∂_l Γ^i_{jk} + Γ^i_{mk}Γ^m_{jl} - Γ^i_{ml}Γ^m_{jk}
                        R[i,j,k,l] = (
                            gamma[i,j,l,k] - gamma[i,j,k,l] +
                            torch.sum(gamma[i,:,k] * gamma[:,j,l]) -
                            torch.sum(gamma[i,:,l] * gamma[:,j,k])
                        )
        
        # Contract to get Ricci tensor
        ricci = torch.einsum('ijij->ij', R)
        
        return ricci

    def quantum_ricci_flow(
        self,
        point: Tensor,
        time: float,
        dt: float = 0.01,
        steps: int = 100
    ) -> Tensor:
        """Evolve metric under quantum Ricci flow.
        
        Args:
            point: Initial point on manifold
            time: Total evolution time
            dt: Time step size
            steps: Number of integration steps
            
        Returns:
            Evolved point after Ricci flow
        """
        current_point = point
        
        for _ in range(steps):
            # Compute Ricci tensor
            ricci = self.compute_ricci_tensor(current_point)
            
            # Update metric according to Ricci flow equation
            # dg/dt = -2Ric
            current_point = current_point - 2 * dt * torch.einsum(
                'ij,j->i',
                ricci,
                current_point
            )
            
            # Project back to symplectic manifold if needed
            if self.preserve_structure:
                current_point = self._handle_dimension(current_point)
        
        return current_point 